+++ ESI.NET/Models/Universe/IDLookup.cs
@@ -34,9 +34,5 @@ namespace ESI.NET.Models.Universe
 
         [JsonProperty("stations")]
         public List<ResolvedInfo> Stations { get; set; } = new List<ResolvedInfo>();
-
-        [JsonProperty("structures")]
-        public List<ResolvedInfo> Structures { get; set; } = new List<ResolvedInfo>();
-
     }
 }